Allow elements with display:contents to be focused.
This is as concluded (though without a group resolution) in
https://github.com/w3c/csswg-drafts/issues/2632 and as discussed a bit
in https://github.com/whatwg/html/issues/1837 .  Note that this diverges
from Gecko (see https://bugzil.la/1553549, https://bugzil.la/1791648,
and https://github.com/mozilla/standards-positions/issues/772) and
WebKit (see https://bugs.webkit.org/show_bug.cgi?id=255149 and
https://github.com/WebKit/standards-positions/issues/164).

This makes the <slot> element focusable when it has a tabindex.  Note
that this does not match Gecko and WebKit, which also require that its
'display' value is changed to make it focusable.

Note that the <slot> element is still not tabbable in Chrome; this is
https://crbug.com/1428419.  This also does not match Gecko and WebKit,
where it is tabbable whenever it is focusable.  The added test
slot-element-tabbable.tentative.html fails both tests as a result, but
is added anyway.  (Gecko and WebKit pass the display: block test but
fail the default style (display: contents) test.)

The added tests display-contents-focusable-001.html and
slot-element-focusable.tentative.html are partly fixed by this change,
but the style part is still broken due to https://crbug.com/1428420.

Fixed: 1366037
Bug: 1428419, 1428420
Change-Id: I46a8ad3b6442ce07f440c8f6a07210bac305600e