Do not restart browser when there is test type switch
This introduces a new message to allow test runner to switch executor,
instead of restarting the browser altogether.

The TestExecutor and Protocol classes now do not have a logger method
until setup is done. This means they should not log any message inside
their constructor, and this is the case right now.

Bug: 40337