sensors: Reset buffer and reading data in MockSensor.stopReading().
This makes the JS mocks implementation match what we do in
PlatformSensor more closely. This behavior is also set in the spec, as
the readings are supposed to be set to null when the sensor stops
(https://w3c.github.io/sensors/#set-sensor-settings).

The Generic Sensor web tests have been converted to Chrome Driver and
are unaffected by this change, but the Device Orientation tests still
use these mocks. The change does not cause any tests to fail though.

Bug: 1278377
Change-Id: I31d53a657e94630890fe52cd5207b38308265376