[scroll-start] scroll-start-target should align with scroll-snap-align
There is a proposal[1] for a more generic element-level scroll  alignment indicator which would apply to scroll-start-targets and would mean that scroll-snap-align dictates the alignment of
scroll-start-targets.
It is likely that that work will be realized later than
scroll-start-target.
This patch allows scroll-snap-align directly specify the alignment of
scroll-start-targets; this behavior is closer to the end-goal of that proposal.

[1]https://github.com/w3c/csswg-drafts/issues/9012#issuecomment-1749344005

Bug: 1439807
Change-Id: Id6ff1937a806d00c3ac0326d53d379d1c39b5897