Map colors into gamut limits before rendering.
When colors are converted to sRGB to render onto the display, make sure
that they are within sRGB gamut limits.

Gamut mapping is implemented according to:
https://drafts.csswg.org/css-color-4/#gamut-mapping

The color-mix-non-srgb-001 test is checking the expected result in
sRGB, which happens to be out of gamut limits, but because the test
is for color-mix and not gamut mapping, I changed the expected
results to the color space of the mix.

Differential Revision: https://phabricator.services.mozilla.com/D191083

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1847503
gecko-commit: 667d97f09248c0e83d02862b3a8cbbb2762eac93
gecko-reviewers: layout-reviewers, emilio