Change expectation for invalid/missing shadowRootMode IDL to null
Per the spec discussion:

https://github.com/whatwg/html/pull/5465#discussion_r1360590511

this changes the expectation for the tests to be:

document.createElement("template").shadowRootMode === ""

rather than === null.

Note that this CL does not change Chromium's behavior, and as such
I cannot check if there are other tests that implicitly also
do a similar check. More tests might need to change once I (or
someone) implements this new behavior in a browser.

Bug: 1379513
Change-Id: I0ea1371d612d524fdecc287f3f1790d601c64982