Add CSS filter string support as parameter for beginLayer
This change adds support for using a CSS filter string as the input
parameter for beginLayer. the IDL definition for CanvasFilterInput was
modified to allow the use of a DOMString as the filter when beginLayer
is called.

A new method in CanvasFilterOperationResolver was added for creating the
filter operations for this type of input, since it requires access to
the font value (for units that depend on the font) and potentially to a
style resolution host if possible.

WPT tests were added to verify the behavior of these filters and their
interaction when nested within a layer with a filter that uses a
dictionary.

Bug: 1396372
Change-Id: I0926477f36d4886c467666adf4eb4c61c3a5c33f