Introduce reserved.top_navigation_start/commit.
Automatic beacons currently have one supported beacon type:
reserved.top_navigation. This fires when a navigation commits. There's a
need to expand this to support beacons both at navigation start and
navigation commit. This CL deprecates reserved.top_navigation in favor
of 2 new beacons: reserved.top_navigation_start and
reserved.top_navigation_commit. The deprecation is done in order to have
a more consistent naming convention.

To support multiple automatic beacon types, each beacon type gets its
own automatic beacon data. This is stored in a hash map in the fenced
frame config, using the event type as the key and the
AutomaticBeaconInfo struct as the value.

This CL refactors automatic beacon tests to support tracking multiple
messages. The automatic-beacon-store.py endpoint uses a hash of the
message to store a beacon, and then the test passes in the message
string to be hashed when checking for the existence of the message.

Change-Id: I268f5c75b8ed754cb11e8c176252d124a8270b69