view-transition: Move pagereveal tests to WPT
The `pagereveal` event will be specified and shipped separately from
view transitions so move tests into WPT. Since the spec is still WIP
these tests are marked tentative for now.

Tests that were specifically dealing with usage of the `viewTransition`
attribute and its behavior in relation to a view transition were left in
the wpt_internal directory for view-transitions since those seems
specific to that feature.

Bug: 1466250
Change-Id: I218f5ed18bbbf0a1c176fbf562aa5ae939a3adcc