[CompositeClipPathAnimations] Make composite CC anim tests use steps()
This change is part of an ongoing effort to increase test coverage to increase confidence in a potential change the the method of interpolation in clip paths.

Steps() is used here with careful timing to ensure the composited animation is created in a state just before its first step to ensure that the main thread has not rendered the animation in the expected state, and a screenshot captured just after. This approach helps in accurately testing the composited animation rendering with precise expected values.

Bug: 1447045

Change-Id: Ia1862af04d4c895331a1e642ebe4ca4379771aa9